+++ client/src/pages/Home/components/index.ts
@@ -1,1 +1,2 @@
 export { default as ProductList } from './ProductList';
+export { default as Product } from './Product';
