+++ client/src/pages/index.ts
@@ -1,3 +1,3 @@
-export { default as ProductList } from './ProductList';
+export { default as Home } from './Home';
 export { default as OrderList } from './OrderList';
 export { default as Cart } from './Cart';
